+++ frontend/ui/lib/widgets/create_trip_widget.dart
@@ -1,9 +1,6 @@
-import 'dart:async';
 import 'package:flutter/material.dart';
-import 'package:flutter/services.dart';
-import 'package:tripmeout/services/trip_service.dart';
 import 'package:tripmeout/model/trip.dart';
-import 'package:tripmeout/router/router.dart';
+import 'package:tripmeout/services/trip_service.dart';
 import 'package:tripmeout/widgets/autocomplete_text_field_widget.dart';
 
 //TODO: Add loading screen after to the Create Trip Widget
